Improve representability in jacobian epsilon
Replace abs(y) * sqrt(U) with a term acquired with numpy functions. This guarantees representability in division, not just addition. It also potentially removes a unwanted scalar term for epsilon.

Note that this actually removes the need to have a second term. As this choice guarantees that epsilon is non-zero. When y is zero it yields an epsilon equal 2**(exp_min). The second term is now used avoid unnecessarily small epsilon and should probably be based in the tolerance. Quantitative testing might show this. Alternatively, a minimum chosen similar to sigma_0 in CVode.